app/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
requests
google-adk
google-genai
google-cloud-aiplatform
qdrant-client>=1.9.0